+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "teamprompts-app",
-  "version": "0.16.8",
+  "version": "0.16.9",
   "description": "TeamPrompts - Help your team get on the same page.",
   "private": true,
   "dependencies": {
