+++ src/containers/App/Heading.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Heading({ title }) {
+  return (
+    <h1 className="font-bold mb-8 mt-2 text-4xl text-gray-900">{title}</h1>
+  );
+}
+
+export default Heading;
